\name{Compare-methods}
\docType{methods}
\alias{Compare-methods}
\title{ Methods for Function Compare }
\description{
  Methods for function \code{Compare}
}
\section{Methods}{
  \describe{
    \item{\code{signature(e1 = "transfer", e2 = "numeric")}}{
      Compares a \code{transfer} object with a \code{numeric} object.
}
}}
\keyword{methods}

