% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getParams}
\alias{getParams}
\title{Extract Transfer and Persistence Parameters I}
\usage{
getParams(tferObj)
}
\arguments{
\item{tferObj}{An object of class \code{transfer}}
}
\value{
\code{getParams} returns a list of input parameters and their corresponding values.
}
\description{
Displays input parameters and arguments passed to \code{\link{transfer}}.
}
\details{
\code{getParams} is one of the two accessor functions for a \code{transfer}
object.
}
\examples{

library(tfer)

y = transfer()

getParams(y)

}
\seealso{
\code{\link{transfer}}
}
\author{
TingYu Huang
}
