% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{getValues}
\alias{getValues}
\title{Extract Transfer Values
n
\code{getValues} is a accessor function which returns the number of recovered
glass fragments generated by \code{\link{transfer}}.}
\usage{
getValues(tferObj)
}
\arguments{
\item{tferObj}{An object of class \code{tfer}}
}
\value{
\code{values} returns a numeric vector of random variates.
}
\description{
Extract Transfer Values
n
\code{getValues} is a accessor function which returns the number of recovered
glass fragments generated by \code{\link{transfer}}.
}
\examples{

library(tfer)
y = transfer()
getValues(y)

}
\seealso{
\code{\link{transfer}}
}
\author{
TingYu Huang and James Curran
}
