\name{parameters-methods}
\docType{methods}
\alias{parameters-methods}
\title{ Methods for Function parameters }
\description{
  Methods for function \code{parameters}
}
\section{Methods}{
\describe{
  \item{\code{signature(object = "transfer")}}{
    This is currently the only method for generic function
    \code{parameters}. It displays the input parameters and arguments
    provided by the user as a string.
}
}}
\keyword{methods}

