% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.transfer.R
\name{plot.tfer}
\alias{plot.tfer}
\title{plot method for objects of transfer class}
\usage{
\method{plot}{tfer}(
  x,
  ptype = c("density", "freq", "hist"),
  xlab = "n",
  main = "",
  col = "red",
  ...
)
}
\arguments{
\item{x}{an object of class \code{transfer}}

\item{ptype, }{one of \code{"density"}, \code{"freq"}, or \code{"hist"}. \code{"density"} will give a barplot with probability
on the y-axis, \code{"frequency"} will give a barplot with frequencies (counts) on the y-axis, and \code{"hist"} will
produce a historgram with frequency (counts) on the y-axis. One-letter versions will also work, i.e. \code{"d"}, \code{"f"} and
\code{"h"}. The original \code{0}, \code{1}, \code{2} will also work, but this usage is deprecated and will produce a warning.}

\item{xlab}{the x-axis label, by default "n"}

\item{main}{the plot title, empty by default}

\item{col}{the colour of the bars in the plot, by default "red"}

\item{\dots}{any other arguments to be passed to \code{barplot} or \code{histogram}}
}
\description{
plot method for objects of transfer class
}
