% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.transfer.R
\name{print.transfer}
\alias{print.transfer}
\title{\code{print} method for transfer objects}
\usage{
\method{print}{transfer}(x, ...)
}
\arguments{
\item{x}{an object of class transfer}

\item{\dots}{included for consistency but not used}
}
\description{
Prints a summary of the simulation input parameters
}
\keyword{methods}
\keyword{print}
