\name{show-methods}
\docType{methods}
\alias{show-methods}
\title{ Methods for Function show }
\description{
  Methods for function \code{show}
}
\section{Methods}{
\describe{
  \item{\code{signature(object = "transfer")}}{
    Displays input parameters and simulated values as a list.
  }
}}
\keyword{methods}
\keyword{show}
