\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\title{ Methods for Function summary }
\description{
  Methods for function \code{summary}
}
\section{Methods}{
\describe{
  \item{\code{signature(object = "ANY")}}{
    \code{summary} methods for any \R objects.
  }
  
  \item{\code{signature(object = "transfer")}}{
    Returns summary statistics of a \code{transfer} object.
}
}}
\keyword{methods}
\keyword{summary}
