% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.transfer.R
\name{summary.transfer}
\alias{summary.transfer}
\title{\code{summary} method for transfer objects}
\usage{
\method{summary}{transfer}(object, ...)
}
\arguments{
\item{object}{an object of class transfer}

\item{\dots}{extra arguments passed to \code{\link{summary.default}}}
}
\value{
A \code{list} with three elements is returned invisibly: 
\describe{
  \item{parameters}{list containing all the simulation parameters}
  \item{values}{a numeric vector of the simulated values}
  \item{probability}{a named numeric vector giving the probability of recovering 0, 1, 2, \ldots fragments}
}
}
\description{
Prints a summary of the simulation input parameters
}
\keyword{methods}
\keyword{summary}
