\name{values}
\alias{values}
\title{
  Extract Transfer Values
}
\description{
  \code{values} is a generic function which returns the number of
  recovered glass fragments generated by \code{\link{transfer}}.
}
\usage{
values(object)
}
\arguments{
  \item{object}{An object of class \code{transfer}}
}
\value{
  \code{values} returns a numeric vector of random variates.
}
\author{
TingYu Huang
}
\seealso{
\code{\link{transfer}}
}
\examples{
library(tfer)
y = transfer()
values(y)
}
