\name{[-methods}
\docType{methods}
\alias{[-methods}
\title{ Methods for Function [ }
\description{
  Methods for function \code{[} 
}
\section{Methods}{
\describe{

\item{\code{signature(x = "transfer", i = "ANY", j = "missing", drop = "missing")}}{
Extracts elements from \code{values(x)}.
}
}}
\keyword{methods}

